Remove response stub from LoaderFunctionArgs/ActionFunctionArgs
