+++ .changeset/fresh-pigs-leave.md
@@ -1,0 +1,6 @@
+---
+"@remix-run/server-runtime": patch
+---
+
+Remove `response` stub from `LoaderFunctionArgs`/`ActionFunctionArgs`
+ - Instead, you will need to use `defineLaoder`/`defineAction` with single fetch to gain type access to the `response` stub
